Backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
scikit-learn